Code refactoring

+++ src/main/java/kg/alex/spt/utils/FormattedFilterTable.java
@@ -27,7 +27,7 @@ public class FormattedFilterTable extends FilterTable {
 
             if (property.getValue() != null) // Format a decimal value for a specific locale
             {
-                return Settings.dFormat.format((Double) property.getValue());
+                return Settings.dFormat.format(property.getValue());
             }
         }
         return super.formatPropertyValue(rowId, colId, property);
